+++ src/App.tsx
@@ -105,8 +105,14 @@ const components: DemoComponent[] = [
           "observations",
           "timeline",
           "observations-outside",
+          "care-team",
+          "documents",
+          "immunizations",
         ]}
         title="ZAP"
+        timelineProps={{ enableFQS: true }}
+        observationsProps={{ enableFQS: true }}
+        immunizationsProps={{ enableFQS: true }}
       />
     ),
   },