+++ src/components/content/medications/patient-medications-outside.tsx
@@ -3,19 +3,21 @@ import { useAddMedicationForm } from "./helpers/add-new-med-drawer";
 import { medicationFilters } from "./helpers/filters";
 import { PatientMedicationsBase } from "./helpers/patient-medications-base";
 import { useToggleArchive } from "../hooks/use-toggle-archive";
+import { getDateRangeView } from "../resource/helpers/view-date-range";
 import { withErrorBoundary } from "@/components/core/error-boundary";
 import { RowActionsProps } from "@/components/core/table/table";
 import { MedicationStatementModel } from "@/fhir/models";
 import { useQueryAllPatientMedications } from "@/hooks/use-medications";
 import { useBaseTranslations } from "@/i18n";
 import { Spinner } from "@/index";
-import { QUERY_KEY_OTHER_PROVIDER_MEDICATIONS } from "@/utils/query-keys";
+import { QUERY_KEY_BASIC, QUERY_KEY_OTHER_PROVIDER_MEDICATIONS } from "@/utils/query-keys";
 
 export type PatientMedicationsOutsideProps = {
   className?: string;
   onOpenHistoryDrawer?: () => void;
   onAddToRecord?: (record: MedicationStatementModel) => void;
   readOnly?: boolean;
+  enableFQS?: boolean;
 };
 
 const PatientMedicationsOutsideComponent = ({
@@ -23,9 +25,12 @@ const PatientMedicationsOutsideComponent = ({
   onAddToRecord,
   readOnly = false,
   onOpenHistoryDrawer,
+  enableFQS = false,
 }: PatientMedicationsOutsideProps) => {
-  const { otherProviderMedications, isLoading } = useQueryAllPatientMedications();
+  const { otherProviderMedications, isLoading } = useQueryAllPatientMedications(enableFQS);
   const rowActions = useMemo(() => getRowActions({ onAddToRecord }), [onAddToRecord]);
+  const { viewOptions, defaultView } =
+    getDateRangeView<MedicationStatementModel>("lastActivityDate");
 
   return (
     <PatientMedicationsBase
@@ -33,6 +38,8 @@ const PatientMedicationsOutsideComponent = ({
       query={{ data: otherProviderMedications, isLoading }}
       filters={medicationFilters(otherProviderMedications, true)}
       rowActions={readOnly ? undefined : rowActions}
+      views={viewOptions}
+      defaultView={defaultView}
       onOpenHistoryDrawer={onOpenHistoryDrawer}
     />
   );
@@ -59,7 +66,8 @@ const RowActions = ({ record, onAddToRecord }: RowActionsProps2) => {
   const showAddMedicationForm = useAddMedicationForm();
   const { isLoading, toggleArchive } = useToggleArchive(
     record,
-    QUERY_KEY_OTHER_PROVIDER_MEDICATIONS
+    QUERY_KEY_OTHER_PROVIDER_MEDICATIONS,
+    QUERY_KEY_BASIC
   );
   const archiveLabel = record.isArchived ? t("resourceTable.restore") : t("resourceTable.dismiss");