+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # ctw-component-library
 
+## 1.13.1
+
+### Patch Changes
+
+- fa1fb932: Fix bug when trying to save or update resources fetched from FQS.
+
+## 1.13.0
+
+### Minor Changes
+
+- 4ac4c517: Add view filtering by date for timeline, documents, and outside medication components.
+
 ## 1.12.0
 
 ### Minor Changes